test: move confidence_aware_eql tests to the root test folder
Relocate the tests from inside the package to
test/experiments_test/confidence_aware_eql/, matching the repo's test layout,
and update the cross-test import accordingly.
